refactor: group withoutSequencing tests
